Finished Basic CRUD Actions for Products Table
Create, show, index, delete, and update are now working for
the Products table. Changed the requirements for the Product entries
so that each product doesn't require a user to do CRUD actions. This
allows us to give read access to anyone on the website so products
can be viewed without signing in.