Reworked Cart Model and Show Request
Created another entry in the Cart model called 'item' to
display the cart item information. This is useful to that the
product ids can be isolated from the items list in case queries
need to be made without traversing through the items array. The
show request was modified but the same problem persists, where the
product information will not be sent as a response after the show
request is complete.