Changed Cart Model, User/Cart Controller
Made everything on the Cart model 'required: false'. Added a cartId
field in the user model to store their cart id. Changed cart update
request back to the default function, data will be modified via
the front end and then sent to the update function in JSON with changes
already added to it.